Create README.md
This script defines a new flight mode called "STABLE_ALTITUDE" that uses a PID controller to maintain the drone's upright position as well as a target altitude. The script creates four PID controllers for roll, pitch, yaw, and altitude and sets them to use gyro, orientation, and barometer data for input. In the modeStableAltitude function, the script gets the current roll, pitch,